requirements.txt
# Firebase
firebase-admin>=6.4.0

# Environment
python-dotenv>=1.0.0

# RL & Simulation (for future phases)
gymnasium>=0.29.0
numpy>=1.26.0
pandas>=2.1.0

# LLM / Ollama
requests>=2.31.0

# Utilities
colorama>=0.4.6
tabulate>=0.9.0
